% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dbacktriannorm}
\alias{dbacktriannorm}
\title{author Mav}
\usage{
dbacktriannorm(x, mu = 0, sigma = 1, m = 0.5)
}
\arguments{
\item{x}{a double vector}

\item{mu}{a double number}

\item{sigma}{a double number}

\item{m}{a double number}
}
\value{
a double vector with the same lenth of x
}
\description{
Evaludate the probabiliy density function of a Backtrain Gaussian
distribution
}
\details{
Evaludate the probabiliy density function of a Backtrain
at specified values
}
\examples{
dbacktriannorm(1.0, mu = 0, sigma = 1, m = 0.95)
}

