% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rbacktriannorm}
\alias{rbacktriannorm}
\title{author Maverick}
\usage{
rbacktriannorm(n, mu = 0, sigma = 1, m = 0.5)
}
\arguments{
\item{n}{an integer}

\item{mu}{a double number}

\item{sigma}{a double number}

\item{m}{a double number}
}
\value{
a double vector of length n
}
\description{
Generate random sample from Backtrain Gaussian distribution
}
\details{
Generate random sample from Backtrain Gaussian distribution
}
\examples{
rbacktriannorm(20, mu = 0, sigma = 1, m = 0.95)
}

